+++ codex-rs/core/src/lib.rs
@@ -66,6 +66,7 @@ pub mod shell;
 pub mod spawn;
 pub mod terminal;
 mod tool_apply_patch;
+mod tools;
 pub mod turn_diff_tracker;
 pub use rollout::ARCHIVED_SESSIONS_SUBDIR;
 pub use rollout::RolloutRecorder;
