+++ codex-rs/core/src/tools/handlers/mod.rs
@@ -1,0 +1,15 @@
+mod apply_patch;
+mod exec_stream;
+mod mcp;
+mod plan;
+mod shell;
+mod unified_exec;
+mod view_image;
+
+pub use apply_patch::ApplyPatchHandler;
+pub use exec_stream::ExecStreamHandler;
+pub use mcp::McpHandler;
+pub use plan::PlanHandler;
+pub use shell::ShellHandler;
+pub use unified_exec::UnifiedExecHandler;
+pub use view_image::ViewImageHandler;
