+++ codex-rs/core/src/tools/mod.rs
@@ -1,0 +1,7 @@
+pub mod context;
+pub(crate) mod handlers;
+pub mod registry;
+pub mod router;
+pub mod spec;
+
+pub use router::Router;
